+++ src/sentry/workflow_engine/tasks/__init__.py
@@ -1,1 +1,2 @@
+from .delayed_workflows import *  # noqa
 from .workflows import *  # noqa
